@@ -1,6 +1,6 @@
 {
   "name": "reactivestates",
-  "version": "1.0.4",
+  "version": "2.0.0",
   "main": "dist/index.js",
   "typings": "dist/index.d.ts",
   "files": [
@@ -9,7 +9,7 @@
   "dependencies": {
     "es6-shim": "^0.35.3",
     "lodash": "^4.17.5",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.0.0",
     "ts-lint": "^4.5.1"
   },
   "devDependencies": {
@@ -25,11 +25,11 @@
     "tslint": "^5.9.1",
     "typescript": "^2.7.1",
     "typings": "2.0.0",
-    "webpack": "^1.14.0",
-    "webpack-dev-server": "^1.16.2"
+    "webpack": "^4.8.3",
+    "webpack-dev-server": "^3.1.4"
   },
   "peerDependencies": {
-    "rxjs": "5.5.6"
+    "rxjs": "^6.0.0"
   },
   "scripts": {
     "clean": "del-cli dist",
