@@ -1,5 +1,5 @@
 import {skip} from "rxjs/operators";
-import {Subject} from "rxjs/Subject";
+import {Subject} from "rxjs";
 import {observableToState, State} from "./State";
 
 describe("State", function () {
